[EI-185] [web] chore: Update import paths for actual indicators value view model
