+++ web/app/business/hooks/indicator/use-live-indicators-value-view-model.hook.ts
@@ -1,6 +1,6 @@
 import { useMemo } from 'react';
 import { useFetchLiveIndicatorsValue } from '../../../store/querys/numerical-guidance/indicator.query';
-import { convertLiveIndicatorsValueViewModel } from '../../services/view-model/indicators-value-view-model.service';
+import { convertLiveIndicatorsValueViewModel } from '../../services/view-model/actual-indicators-value-view-model.service';
 import { useSelectedIndicatorBoardMetadata } from '../indicator-board-metedata/use-selected-indicator-board-metadata-view-model.hook';
 import { useWorkspaceStore } from '@/app/store/stores/numerical-guidance/workspace.store';
