+++ web/app/business/services/chart/indicator-formatter.service.ts
@@ -1,4 +1,4 @@
-import { UnitType, IndicatorValue, FormattedItem } from '../view-model/indicators-value-view-model.service';
+import { UnitType, IndicatorValue, FormattedItem } from '../view-model/actual-indicators-value-view-model.service';
 import { utcFormat, utcParse } from 'd3-time-format';
 
 export const parseTime = utcParse('%Y%m%d');
