+++ web/app/business/services/view-model/custom-forecast-indicator-value-view-model.service.ts
@@ -1,0 +1,1 @@
+// export class CustomForecastIndicatorValue
