+++ web/app/ui/components/view/molocule/advanced-multi-line-chart/advanced-multi-line-chart.stories.tsx
@@ -1,7 +1,7 @@
 import type { Meta, StoryObj } from '@storybook/react';
 
 import { indicatorsValueMockData } from '@/app/mocks/mock-data/indicators-value.mock';
-import { convertLiveIndicatorsValueViewModel } from '@/app/business/services/view-model/indicators-value-view-model.service';
+import { convertLiveIndicatorsValueViewModel } from '@/app/business/services/view-model/actual-indicators-value-view-model.service';
 import AdvancedMultiLineChart from './advanced-multi-line-chart';
 import { IndicatorFormatter } from '@/app/business/services/chart/indicator-formatter.service';
 import { Card } from '@tremor/react';
